+++ CHANGES.md
@@ -4,6 +4,7 @@
 Changes in 10.6.1
 -----------------
 - PostgreSQL Fix Class Rank float comparison issue: do NOT use double precision type (inexact), use numeric (exact) in rosariosis.sql, thanks to @fatahou
+- MySQL Fix Class Rank float comparison issue: do NOT use double precision type (inexact), use numeric(22,16) (exact) in rosariosis_mysql.sql, thanks to @fatahou
 Changes in 10.6
 ---------------
 - Add Pashto to right-to-left languages in Warehouse.php, Inputs.php & PDF.php